refactoring of AllEventsList so it can be used for AllEventsPage and PublicOrgPage
