+++ src/features/organizations/l10n/messageIds.ts
@@ -3,26 +3,6 @@ import { ReactElement } from 'react';
 import { m, makeMessages } from 'core/i18n';
 
 export default makeMessages('feat.organizations', {
-  allEventsList: {
-    emptyList: {
-      message: m('Could not find any events'),
-      removeFiltersButton: m('Clear filters'),
-    },
-    filterButtonLabels: {
-      organizations: m<{ numOrgs: number }>(
-        '{numOrgs, plural,=0 {Organizations} =1 {1 organization} other {# organizations}}'
-      ),
-      thisWeek: m('This week'),
-      today: m('Today'),
-      tomorrow: m('Tomorrow'),
-    },
-  },
-  authDialog: {
-    cancelButton: m('Cancel'),
-    content: m('You need a Zetkin account to sign up for events.'),
-    label: m('Sign up'),
-    loginButton: m('Log in & come back'),
-  },
   gen3: {
     description: m(
       'This is the new (generation 3) organizer web app. If you are used to the old one, you will find lots of new features and an improved user interface here. But if you want, you can still use the old organizer app for a while longer.'
